synq: relocate pass-result types from lsp/ into semantic::analysis
First step of moving SQL query analysis into a composable pass
pipeline. These types are semantic-analyzer outputs that LspObserver
has been hoarding in lsp/; they belong in semantic/.

- new semantic::analysis module owns StoredToken, StoredComment,
  SemanticToken, ResolvedSymbol, Resolution, SymbolIdentity,
  DefinitionLocation/Result, and DocumentAnalysisData + its
  resolution_at / references_matching / definition_at /
  semantic_tokens query methods
- lsp/analysis_data.rs trimmed to LSP-only glue: LspObserver,
  ExternalDefinitions, CompletionContext/Info
- imports updated in lsp/{host,hover_service,document_store,
  completion,completion_service}
- completion_service::build_completion_items now takes
  CompletionInfo by ref (clippy pedantic; same subsystem)
- module is feature-gated on "lsp" for now; widens in step 2
  when non-LSP passes start consuming it

No behavior change.
