+++ syntaqlite/src/lsp/analysis_data.rs
@@ -1,48 +1,30 @@
 // Copyright 2025 The syntaqlite Authors. All rights reserved.
 // Licensed under the Apache License, Version 2.0.
 
-//! LSP-local storage captured during a semantic-analyzer pass.
+//! LSP-side glue for capturing semantic-analyzer events.
 //!
-//! The semantic layer emits events via [`AnalysisObserver`]; this module
-//! defines the concrete storage (tokens, comments, resolutions, definition
-//! offsets) plus an observer implementation that fills it. Nothing here is
-//! visible to the semantic layer.
+//! The analysis output types themselves live in [`crate::semantic::analysis`].
+//! This module holds:
+//!
+//! - [`LspObserver`] — an [`AnalysisObserver`] impl that fills a
+//!   [`DocumentAnalysisData`] during a pass.
+//! - [`ExternalDefinitions`] — a cross-file definition-site registry consulted
+//!   by the observer to correlate resolutions with schema files.
+//! - [`CompletionInfo`] / [`CompletionContext`] — the completion-probe output
+//!   consumed by the LSP completion service.
 
 use std::collections::HashMap;
 
 use syntaqlite_syntax::ParserTokenFlags;
-use syntaqlite_syntax::any::{AnyTokenType, TokenCategory};
+use syntaqlite_syntax::any::AnyTokenType;
 use syntaqlite_syntax::source::{DocLen, DocOffset, DocRange};
 
-use crate::dialect::AnyDialect;
+use crate::semantic::analysis::{
+    DefinitionLocation, DocumentAnalysisData, Resolution, ResolvedSymbol, StoredComment,
+    StoredToken,
+};
 use crate::semantic::catalog::{AritySpec, FunctionCategory};
 use crate::semantic::observer::AnalysisObserver;
-
-// ── Token positions ──────────────────────────────────────────────────────────
-
-/// A parser token observed during analysis.
-#[derive(Debug, Clone)]
-pub(crate) struct StoredToken {
-    pub(crate) offset: DocOffset,
-    pub(crate) length: DocLen,
-    pub(crate) token_type: AnyTokenType,
-    pub(crate) flags: ParserTokenFlags,
-}
-
-/// A comment observed during analysis.
-#[derive(Debug, Clone)]
-pub(crate) struct StoredComment {
-    pub(crate) offset: DocOffset,
-    pub(crate) length: DocLen,
-}
-
-/// A token classified for editor syntax highlighting.
-#[derive(Debug, Clone)]
-pub(crate) struct SemanticToken {
-    pub(crate) offset: DocOffset,
-    pub(crate) length: DocLen,
-    pub(crate) category: TokenCategory,
-}
 
 // ── Completion ───────────────────────────────────────────────────────────────
 
@@ -71,96 +53,6 @@ pub(crate) struct CompletionInfo {
     pub(crate) tokens: Vec<AnyTokenType>,
     pub(crate) context: CompletionContext,
     pub(crate) qualifier: Option<String>,
-}
-
-// ── Resolved symbols ─────────────────────────────────────────────────────────
-
-/// A definition site that a reference points to.
-#[derive(Debug, Clone)]
-pub(crate) struct DefinitionLocation {
-    pub(crate) range: DocRange,
-    /// `Some` when the definition lives in another file (external schema).
-    pub(crate) file_uri: Option<String>,
-}
-
-/// Result of a go-to-definition lookup.
-#[derive(Debug, Clone)]
-pub(crate) struct DefinitionResult {
-    pub(crate) origin: DocRange,
-    pub(crate) target: DefinitionLocation,
-}
-
-/// A symbol resolution recorded during the validation pass.
-#[derive(Debug, Clone)]
-pub(crate) enum ResolvedSymbol {
-    Table {
-        name: String,
-        columns: Option<Vec<String>>,
-        definition: Option<DefinitionLocation>,
-    },
-    Column {
-        column: String,
-        table: String,
-        all_columns: Vec<String>,
-        definition: Option<DefinitionLocation>,
-    },
-    Function {
-        category: String,
-        arities: Vec<String>,
-    },
-}
-
-#[derive(Debug, Clone)]
-pub(crate) struct Resolution {
-    pub(crate) range: DocRange,
-    pub(crate) symbol: ResolvedSymbol,
-}
-
-/// Identity of a symbol for matching across resolutions.
-#[derive(Debug)]
-pub(crate) enum SymbolIdentity {
-    Table(String),
-    Column { table: String, column: String },
-}
-
-impl SymbolIdentity {
-    pub(crate) fn from_resolved(sym: &ResolvedSymbol) -> Option<Self> {
-        match sym {
-            ResolvedSymbol::Table { name, .. } => {
-                Some(SymbolIdentity::Table(name.to_ascii_lowercase()))
-            }
-            ResolvedSymbol::Column { column, table, .. } => Some(SymbolIdentity::Column {
-                table: table.to_ascii_lowercase(),
-                column: column.to_ascii_lowercase(),
-            }),
-            ResolvedSymbol::Function { .. } => None,
-        }
-    }
-
-    fn matches(&self, sym: &ResolvedSymbol) -> bool {
-        match (self, sym) {
-            (SymbolIdentity::Table(name), ResolvedSymbol::Table { name: n, .. }) => {
-                n.eq_ignore_ascii_case(name)
-            }
-            (
-                SymbolIdentity::Column { table, column },
-                ResolvedSymbol::Column {
-                    table: t,
-                    column: c,
-                    ..
-                },
-            ) => t.eq_ignore_ascii_case(table) && c.eq_ignore_ascii_case(column),
-            _ => false,
-        }
-    }
-
-    /// Key into `definition_offsets` for this symbol.
-    pub(crate) fn definition_key(&self) -> String {
-        match self {
-            SymbolIdentity::Table(name) => name.clone(),
-            SymbolIdentity::Column { table, column } => format!("{table}.{column}"),
-        }
-    }
 }
 
 // ── External definition registry ─────────────────────────────────────────────
@@ -229,75 +121,6 @@ impl ExternalDefinitions {
             column.to_ascii_lowercase()
         );
         self.sites.get(&key)
-    }
-}
-
-// ── DocumentAnalysisData ─────────────────────────────────────────────────────
-
-/// All the data an editor needs that was captured during a single analysis
-/// pass. Populated by [`LspObserver`].
-#[derive(Debug, Default)]
-pub(crate) struct DocumentAnalysisData {
-    pub(crate) tokens: Vec<StoredToken>,
-    pub(crate) comments: Vec<StoredComment>,
-    pub(crate) resolutions: Vec<Resolution>,
-    /// Maps `lowercase(name)` → `DocRange` for same-file definition sites.
-    /// Keys for columns look like `"table.column"` (lowercased).
-    pub(crate) definition_offsets: HashMap<String, DocRange>,
-}
-
-impl DocumentAnalysisData {
-    pub(crate) fn semantic_tokens(&self, dialect: &AnyDialect) -> Vec<SemanticToken> {
-        let mut out = Vec::new();
-        for t in &self.tokens {
-            let cat = dialect.classify_token(t.token_type, t.flags);
-            if cat != TokenCategory::Other {
-                out.push(SemanticToken {
-                    offset: t.offset,
-                    length: t.length,
-                    category: cat,
-                });
-            }
-        }
-        for c in &self.comments {
-            out.push(SemanticToken {
-                offset: c.offset,
-                length: c.length,
-                category: TokenCategory::Comment,
-            });
-        }
-        out.sort_by_key(|t| t.offset);
-        out
-    }
-
-    /// The resolution whose span contains `offset`, if any.
-    pub(crate) fn resolution_at(&self, offset: DocOffset) -> Option<&Resolution> {
-        self.resolutions
-            .iter()
-            .find(|r| offset >= r.range.start && offset < r.range.end)
-    }
-
-    /// Find all resolutions in this document that match the given identity.
-    pub(crate) fn references_matching(&self, kind: &SymbolIdentity) -> Vec<DocRange> {
-        self.resolutions
-            .iter()
-            .filter(|r| kind.matches(&r.symbol))
-            .map(|r| r.range)
-            .collect()
-    }
-
-    /// Go-to-definition target for the resolution at `offset`, if any.
-    pub(crate) fn definition_at(&self, offset: DocOffset) -> Option<DefinitionResult> {
-        self.resolution_at(offset).and_then(|r| match &r.symbol {
-            ResolvedSymbol::Table { definition, .. }
-            | ResolvedSymbol::Column { definition, .. } => {
-                definition.as_ref().map(|d| DefinitionResult {
-                    origin: r.range,
-                    target: d.clone(),
-                })
-            }
-            ResolvedSymbol::Function { .. } => None,
-        })
     }
 }
