+++ syntaqlite/src/lsp/document_store.rs
@@ -10,8 +10,7 @@
 
 use std::collections::HashMap;
 
-use super::analysis_data::DocumentAnalysisData;
-use super::analysis_data::SemanticToken;
+use crate::semantic::analysis::{DocumentAnalysisData, SemanticToken};
 use crate::semantic::diagnostics::Diagnostic;
 
 /// A single open document with lazily populated analysis caches.
